fix: handle tail returns properly
Fixing tail returns for PUC Lua was easy. Simply remember that
return events may be "return" or "tail return" and treat them the same.

LuaJIT was not so easy. The "return" event cannot be trusted, since
tail calls are undetectable. Luckily, you can detect tail returns
by waiting for the active function to change. Afterwards, we can
determine the correct hook offset by computing the stack depth
of debug.traceback() and comparing it to the initial stack depth
of the debug hook's first event.

For recursive functions in LuaJIT, we can't detect tail returns
since the active function won't change. So we need to avoid incrementing
the hook offset for these recursive calls. Likewise, we then need
to avoid decrementing the hook offset when these recursive calls
return. In this way, recursive function calls are treated as a
single function (even if tail calls are not used).